+++ mixly_arduino/blockly/generators/arduino/esp32/Handbit.js
@@ -1,9 +1,7 @@
 'use strict';
 
 goog.provide('Blockly.Arduino.Handbit');
-
 goog.require('Blockly.Arduino');
-
 Blockly.Arduino.oled_init = function () {
   Blockly.Arduino.definitions_['include_U8g2lib'] = '#include <U8g2lib.h>';
   Blockly.Arduino.definitions_['var_declare_U8G2'] = 'U8G2_SH1106_128X64_NONAME_F_HW_I2C u8g2(U8G2_R0, SCL, SDA, U8X8_PIN_NONE);';
@@ -195,13 +193,9 @@ Blockly.Arduino.handbit_button_is_pressed = function () {
   return [code, Blockly.Arduino.ORDER_ATOMIC];
 };
 
-Blockly.Arduino.handbit_light = function () {
-  return ['analogRead(39)', Blockly.Arduino.ORDER_ATOMIC];
-};
+Blockly.Arduino.handbit_light =Blockly.Arduino.sensor_light;
 
-Blockly.Arduino.handbit_sound = function () {
-  return ['analogRead(36)', Blockly.Arduino.ORDER_ATOMIC];
-};
+Blockly.Arduino.handbit_sound =Blockly.Arduino.sensor_sound;
 
 //传感器_重力感应块
 Blockly.Arduino.handbit_MSA300 = function () {
@@ -301,7 +295,7 @@ Blockly.Arduino.handbit_rgb_Brightness = function () {
 };
 
 Blockly.Arduino.handbit_rgb_show = function () {
-  var code = 'rgb_display_17.show();\n';
+  var code = 'rgb_display_17.show();\ndelay(1);\n';
   return code;
 };
 
@@ -346,4 +340,20 @@ Blockly.Arduino.handbit_rgb_rainbow3 = function () {
   else
     var code3 = 'for (int i = 0; i < rgb_display_17.numPixels(); i++)\n {rgb_display_17.setPixelColor(i, Wheel(((i * 256 / rgb_display_17.numPixels()) + ' + rainbow_color + ') & 255));\n}\n';
   return code3;
+};
+
+Blockly.Arduino.board_setup = function () {
+  Blockly.Arduino.definitions_['include_Adafruit_NeoPixel'] = '#include <Adafruit_NeoPixel.h>';
+  Blockly.Arduino.definitions_['var_declare_rgb_display17'] = 'Adafruit_NeoPixel rgb_display_17= Adafruit_NeoPixel(3,17,NEO_GRB + NEO_KHZ800);';
+  Blockly.Arduino.definitions_['include_U8g2lib'] = '#include <U8g2lib.h>';
+  Blockly.Arduino.definitions_['var_declare_U8G2'] = 'U8G2_SH1106_128X64_NONAME_F_HW_I2C u8g2(U8G2_R0, SCL, SDA, U8X8_PIN_NONE);';
+  Blockly.Arduino.definitions_['include_Wire'] = '#include <Wire.h>';
+  Blockly.Arduino.setups_['setup_u8g2.begin'] = "u8g2.begin();";
+    Blockly.Arduino.setups_['setup_rgb_display_begin_17'] = 'rgb_display_17.begin();';
+     Blockly.Arduino.setups_['setup_u8g2.clearDisplay'] = 'u8g2.clearDisplay();';
+      Blockly.Arduino.setups_['setup_RGB_init'] = ' for(int j=0;j<3;j++)\n{for(int i=0;i<3;i++)\nrgb_display_17.setPixelColor(i,0x000000);\nrgb_display_17.show();}\ndigitalWrite(P6,LOW);';
+     
+
+  //return code3;
+
 };