+++ mixly_arduino/blockly/generators/arduino/esp32/actuator.js
@@ -34,26 +34,40 @@ Blockly.Arduino.servo_read_degrees = function() {
   return [code, Blockly.Arduino.ORDER_ATOMIC];
 };
 
-Blockly.Arduino.tone_notes = function() {
-  var code = this.getFieldValue('STAT');
-  return [code, Blockly.Arduino.ORDER_ATOMIC];
-};
-
 Blockly.Arduino.controls_tone=function(){
-    var fre = Blockly.Arduino.valueToCode(this, 'FREQUENCY',Blockly.Arduino.ORDER_ASSIGNMENT) || '0';
-    var channel = Blockly.Arduino.valueToCode(this, 'CHANNEL',Blockly.Arduino.ORDER_ASSIGNMENT) || '0';
+  Blockly.Arduino.definitions_['include_ESP32Tone'] = '#include <ESP32Tone.h>';
+  var dropdown_pin = Blockly.Arduino.valueToCode(this, 'PIN',Blockly.Arduino.ORDER_ATOMIC);
+  var fre = Blockly.Arduino.valueToCode(this, 'FREQUENCY',Blockly.Arduino.ORDER_ASSIGNMENT) || '0';
+  var channel = Blockly.Arduino.valueToCode(this, 'CHANNEL',Blockly.Arduino.ORDER_ASSIGNMENT) || '0';
+  var DELAY_TIME = Blockly.Arduino.valueToCode(this, 'DELAY_TIME',Blockly.Arduino.ORDER_ASSIGNMENT) || '0';
   var code = ""; 
-  code = "ledcWriteTone("+channel+","+fre+");\n";
+  code = " tone("+dropdown_pin+", "+fre+", "+DELAY_TIME+", "+channel+");\n";
   return code;
 };
 
 Blockly.Arduino.controls_notone=function(){
-   var channel = Blockly.Arduino.valueToCode(this, 'CHANNEL',Blockly.Arduino.ORDER_ASSIGNMENT) || '0';
+  var dropdown_pin = Blockly.Arduino.valueToCode(this, 'PIN',Blockly.Arduino.ORDER_ATOMIC);
+  var channel = Blockly.Arduino.valueToCode(this, 'CHANNEL',Blockly.Arduino.ORDER_ASSIGNMENT) || '0';
   var code = ""; 
-  code = "ledcWrite("+channel+",0);\n";
-   return code;
+  code = " noTone("+dropdown_pin+", "+channel+");\n";
+  return code;
+};
+Blockly.Arduino.onboard_tone=function(){
+  Blockly.Arduino.definitions_['include_ESP32Tone'] = '#include <ESP32Tone.h>';
+  var fre = Blockly.Arduino.valueToCode(this, 'FREQUENCY',Blockly.Arduino.ORDER_ASSIGNMENT) || '0';
+  var channel = Blockly.Arduino.valueToCode(this, 'CHANNEL',Blockly.Arduino.ORDER_ASSIGNMENT) || '0';
+  var DELAY_TIME = Blockly.Arduino.valueToCode(this, 'DELAY_TIME',Blockly.Arduino.ORDER_ASSIGNMENT) || '0';
+  var code = ""; 
+  code = " tone(BUZZER, "+fre+", "+DELAY_TIME+", "+channel+");\n";
+  return code;
 };
 
+Blockly.Arduino.onboard_notone=function(){
+ var channel = Blockly.Arduino.valueToCode(this, 'CHANNEL',Blockly.Arduino.ORDER_ASSIGNMENT) || '0';
+ var code = ""; 
+ code = " noTone(BUZZER, "+channel+");\n";
+ return code;
+};
 //执行器-电机转动
 Blockly.Arduino.Mixly_motor = function() {
   var SPEED_PIN = Blockly.Arduino.valueToCode(this, 'PIN1', Blockly.Arduino.ORDER_ATOMIC);
@@ -76,13 +90,11 @@ Blockly.Arduino.HR8833_Motor_Setup = function() {
   var motor_id = Blockly.Arduino.valueToCode(this, 'MOTOR_ID', Blockly.Arduino.ORDER_ATOMIC);
   var pin1 = Blockly.Arduino.valueToCode(this, 'PIN1', Blockly.Arduino.ORDER_ATOMIC);
   var pin2 = Blockly.Arduino.valueToCode(this, 'PIN2', Blockly.Arduino.ORDER_ATOMIC);
-
   Blockly.Arduino.definitions_['HR8833_Motor_Setup_fun'] ='void HR8833_Motor_Setup(int motorID,int pin1,int pin2){//电机初始化 ID=1~4 定义四组电机\n'
-                                                +'    ledcSetup(motorID*2-2, 5000, 8);ledcAttachPin(pin1, motorID*2-2);\n'
-                                                +'    ledcSetup(motorID*2-1, 5000, 8);ledcAttachPin(pin2, motorID*2-1);\n'
-                                                +' }';
+  +'    ledcSetup(motorID*2-2, 5000, 8);ledcAttachPin(pin1, motorID*2-2);\n'
+  +'    ledcSetup(motorID*2-1, 5000, 8);ledcAttachPin(pin2, motorID*2-1);\n'
+  +' }';
   Blockly.Arduino.setups_['motorID_'+motor_id]='HR8833_Motor_Setup('+motor_id+','+pin1+','+pin2+');';
-
   var code =' ';
   return code;
 };
@@ -90,8 +102,6 @@ Blockly.Arduino.HR8833_Motor_Setup = function() {
 Blockly.Arduino.HR8833_Motor_Speed = function() {
   var motor_id = Blockly.Arduino.valueToCode(this, 'MOTOR_ID', Blockly.Arduino.ORDER_ATOMIC);
   var speed = Blockly.Arduino.valueToCode(this, 'SPEED', Blockly.Arduino.ORDER_ATOMIC);
- 
-
   Blockly.Arduino.definitions_['HR8833_Motor_Speed_fun'] ='void HR8833_Motor_Speed(int motorID,int speed){//电机速度设置 ID=1~4,speed=-255~255\n'
   +'if (speed == 0){  \n'
   +'  ledcWrite(motorID*2-2, 0);\n'
@@ -106,8 +116,6 @@ Blockly.Arduino.HR8833_Motor_Speed = function() {
   +'     ledcWrite(motorID*2-1, -speed);\n'
   +'    }\n'
   +'}\n';
-
-
   var code ='HR8833_Motor_Speed('+motor_id+','+speed+'); ';
   return code;
 };